# server/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "SDET School MERN store",
    "description": "SWAGGER API Documentation for Students of SDET School"
  },
  "servers": [
    {
      "url": "https://mern-ecommerce.sdet.school/api"
    },
    {
      "url": "http://localhost:4005/api"
    }
  ],
  "paths": {
...
    "/review/add": {
      "post": {
        "tags": [
          "Review"
        ],
        "summary": "Add review",
        "description": "Add review",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewAdd"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review Add",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewAddResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/review": {
      "parameters": [
        {
          "in": "query",
          "name": "page",
          "description": "Page number",
          "required": true,
          "example": 1
        },
        {
          "in": "query",
          "name": "limit",
          "description": "number or orders on the page",
          "required": true,
          "example": 20
        }
      ],
      "get": {
        "tags": [
          "Review"
        ],
        "summary": "List reviews",
        "description": "List Reviews",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All reviews",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/review/{slug}": {
      "parameters": [
        {
          "description": "Product id",
          "name": "slug",
          "required": true,
          "in": "path"
        }
      ],
      "get": {
        "tags": [
          "Review"
        ],
        "summary": "Get product reviews",
        "description": "Get product reviews",
        "responses": {
          "200": {
            "description": "Get product reviews",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSingleProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/review/{id}": {
      "parameters": [
        {
          "description": "Review id",
          "name": "id",
          "required": true,
          "in": "path"
        }
      ],
      "put": {
        "tags": [
          "Review"
        ],
        "summary": "Update review",
        "description": "Update Review",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewAdd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review updated",
            "content": {
              "application/json":{
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/review/approve/{reviewId}": {
      "parameters": [
        {
          "description": "Review id",
          "name": "reviewId",
          "required": true,
          "in": "path"
        }
      ],
      "put": {
        "tags": [
          "Review"
        ],
        "summary": "Approve review",
        "description": "Approve Review",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Review approved",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/review/reject/{reviewId}": {
      "parameters": [
        {
          "description": "Review id",
          "name": "reviewId",
          "required": true,
          "in": "path"
        }
      ],
      "put": {
        "tags": [
          "Review"
        ],
        "summary": "Reject review",
        "description": "Reject Review",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Review rejected",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/review/delete/{id}": {
      "parameters": [
        {
          "description": "Review id",
          "name": "reviewId",
          "required": true,
          "in": "path"
        }
      ],
      "delete": {
        "tags": [
          "Review"
        ],
        "summary": "Reject review",
        "description": "Reject Review",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Delete review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          }
        }
      }
    },
...
    "/product/add": {
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "Add product",
        "description": "Add product",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/AddProduct"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Add product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/product": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Get all products with full product info",
        "description": "Get all products with full product info",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductFullResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/product/{id}": {
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "description": "productId"
        }
      ],
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Get single products with full product info",
        "description": "Get single products with full product info",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Product info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "summary": "Update single products with full product info",
        "description": "Update single products with full product info",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/product/delete/{id}": {
      "parameters": [
        {
          "in": "path",
          "description": "Product id",
          "name": "id",
          "required": true
        }
      ],
      "delete": {
        "tags": [
          "Product"
        ],
        "summary": "Delete product",
        "description": "Delete product",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Delete product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Unauthorized": {
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          }
        }
      },
      "Error": {
        "properties": {
          "error": {
            "type": "string",
            "example": "Error explanation"
          }
        }
      },
...
      "Success": {
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ProductIdResponse": {
        "properties": {
          "product": {
            "type": "object",
            "properties": {
              "taxable": {
                "type": "boolean"
              },
              "isActive": {
                "type": "boolean"
              },
              "brand": {
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "_id": {
                    "type": "string"
                  }
                }
              },
              "_id": {
                "type": "string"
              },
              "sku": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "quantity": {
                "type": "integer"
              },
              "price": {
                "type": "number"
              },
              "imageUrl": {
                "type": "string"
              },
              "imageKey": {
                "type": "string"
              },
              "created": {
                "type": "string",
                "example": "2022-11-30T01:25:25.197Z"
              },
              "slug": {
                "type": "string"
              },
              "__v": {
                "type": "integer"
              }
            }
          }
        }
      },
      "ProductFullResponse": {
        "type": "object",
        "properties": {
          "taxable": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "brand": {
            "type": "object",
            "properties": {
              "isActive": {
                "type": "boolean"
              },
              "merchant": {
                "type": "object",
                "properties": {
                  "isActive": {
                    "type": "boolean"
                  },
                  "status": {
                    "type": "string",
                    "example": "Approved"
                  },
                  "_id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "business": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "brand": {
                    "type": "string",
                    "description": "Brand name"
                  },
                  "created": {
                    "type": "string",
                    "example": "2022-11-30T01:25:25.197Z"
                  },
                  "__v": {
                    "type": "integer"
                  }
                }
              },
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "created": {
                "type": "string",
                "example": "2022-11-30T01:25:25.197Z"
              },
              "slug": {
                "type": "string"
              },
              "__v": {
                "type": "integer"
              }
            }
          },
          "_id": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "price": {
            "type": "integer"
          },
          "imageUrl": {
            "type": "string"
          },
          "imageKey": {
            "type": "string"
          },
          "created": {
            "type": "string",
            "example": "2022-11-30T01:25:25.197Z"
          },
          "slug": {
            "type": "string"
          },
          "__v": {
            "type": "integer"
          }
        }
      },
...
      "ProductDeleteResponse": {
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "product": {
            "type": "object",
            "properties": {
              "n": {
                "type": "integer"
              },
              "ok": {
                "type": "integer"
              },
              "deletedCount": {
                "type": "integer"
              }
            }
          }
        }
      },
      "AddProduct": {
        "type": "object",
        "required": [
          "description",
          "name",
          "quantity",
          "price",
          "sku"
        ],
        "properties": {
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "integer"
          },
          "quantity": {
            "type": "integer"
          },
          "image": {
            "type": "string",
            "format": "binary"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "taxable": {
            "type": "integer"
          },
          "brand": {
            "type": "string",
            "description": "Brand id"
          }
        }
      },
      "AddProductResponse": {
        "properties": {
          "_id": {
            "type": "string"
          },
          "taxable": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "brand": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "price": {
            "type": "number"
          },
          "imageUrl": {
            "type": "string"
          },
          "imageKey": {
            "type": "string"
          },
          "created": {
            "type": "string",
            "example": "2022-11-30T01:25:25.197Z"
          },
          "slug": {
            "type": "string"
          },
          "__v": {
            "type": "integer"
          }
        }
      },
      "ProductUpdate": {
        "type": "object",
        "properties": {
          "product": {
            "properties": {
              "name": {
                "type": "string"
              },
              "sku": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "quantity": {
                "type": "integer"
              },
              "price": {
                "type": "integer"
              },
              "taxable": {
                "type": "boolean"
              },
              "brand": {
                "type": "array",
                "items": {
                  "type": "string",
                  "description": "Brand id"
                }
              }
            }
          }
        }
      },
...
      "ReviewAdd": {
        "properties": {
          "isRecommended": {
            "type": "string",
            "example": "1",
            "description": "1 ore 0"
          },
          "product": {
            "type": "string",
            "description": "product id"
          },
          "rating": {
            "type": "string",
            "example": "4"
          },
          "review": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "ReviewAddResponse": {
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "review": {
            "type": "object",
            "properties": {
              "product": {
                "type": "string",
                "description": "product id"
              },
              "user": {
                "type": "string",
                "description": "User id"
              },
              "rating": {
                "type": "integer"
              },
              "isRecommended": {
                "type": "boolean"
              },
              "status": {
                "type": "string"
              },
              "_id": {
                "type": "string"
              },
              "review": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "created": {
                "type": "string"
              },
              "__v": {
                "type": "integer"
              }
            }
          }
        }
      },
      "ReviewResponse": {
        "properties": {
          "reviews": {
            "type": "array",
            "items": {
              "properties": {
                "product": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "imageUrl": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    }
                  }
                },
                "user": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    }
                  }
                },
                "rating": {
                  "type": "integer"
                },
                "isRecommended": {
                  "type": "boolean"
                },
                "status": {
                  "type": "string"
                },
                "_id": {
                  "type": "string"
                },
                "review": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "created": {
                  "type": "string"
                },
                "__v": {
                  "type": "integer"
                }
              }
            }
          },
          "totalPages": {
            "type": "integer"
          },
          "currentPage": {
            "type": "integer"
          },
          "count": {
            "type": "integer"
          }
        }
      },
      "ReviewSingleProductResponse": {
        "properties": {
          "reviews": {
            "type": "array",
            "items": {
              "properties": {
                "product": {
                  "type": "string",
                  "description": "Product id"
                },
                "user": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    }
                  }
                },
                "rating": {
                  "type": "integer"
                },
                "isRecommended": {
                  "type": "boolean"
                },
                "status": {
                  "type": "string"
                },
                "_id": {
                  "type": "string"
                },
                "review": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "created": {
                  "type": "string"
                },
                "__v": {
                  "type": "integer"
                }
              }
            }
          }
        }
      },
      "ReviewDeleteResponse": {
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "review": {
            "type": "object",
            "properties": {
              "n": {
                "type": "integer"
              },
              "ok": {
                "type": "integer"
              },
              "deleteCount": {
                "type": "integer"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}